Fix code style.

+++ src/gui/windows/mailwindow.cpp
@@ -253,7 +253,7 @@ void MailWindow::viewPrev(const int id)
             if (it == mMessages.begin())
             {
                 it = mMessages.end();
-                mListBox->setSelected(mMessages.size() - 1);
+                mListBox->setSelected(static_cast<int>(mMessages.size()) - 1);
             }
             else
             {
